+++ core/models.py
@@ -8,11 +8,11 @@ import uuid
 try:
     from .database import db  # type: ignore  # pragma: no cover
 except ImportError:
-    from database import db  # type: ignore
+    from core.database import db  # type: ignore
 try:
     from .enums import ArticleStatus, ArticleVisibility
 except ImportError:  # pragma: no cover - fallback for direct execution
-    from enums import ArticleStatus, ArticleVisibility
+    from core.enums import ArticleStatus, ArticleVisibility
 
 # --- association tables for article visibility ---
 article_extra_celulas = db.Table(
