--- templates/upload.html
@@ -1,30 +1,0 @@
-{% extends "base.html" %}
-
-{% block title %}Upload de Arquivo{% endblock %}
-
-{% block content %}
-  <div class="row justify-content-center">
-    <div class="col-md-6">
-      <div class="card shadow-sm">
-        <div class="card-body">
-          <h3>Upload de Arquivo</h3>
-          {% if current_user.is_authenticated %}
-            <form method="POST" enctype="multipart/form-data">
-              <div class="mb-3">
-                <input type="file" name="file" class="form-control" required>
-              </div>
-              <button type="submit" class="btn btn-primary">Enviar</button>
-            </form>
-          {% else %}
-            <p class="text-muted">Faça login para enviar arquivos.</p>
-          {% endif %}
-          {% if current_user.is_authenticated %}
-            <a href="{{ url_for('logout') }}" class="btn btn-link mt-3">Sair</a>
-          {% else %}
-            <a href="{{ url_for('login') }}" class="btn btn-link mt-3">Entrar</a>
-          {% endif %}
-        </div>
-      </div>
-    </div>
-  </div>
-{% endblock %}
