+++ tests/test_cargo.py
@@ -2,7 +2,7 @@ import pytest
 
 
 from app import app, db
-from models import Cargo, Instituicao, Estabelecimento, Setor, Celula, Funcao, User
+from core.models import Cargo, Instituicao, Estabelecimento, Setor, Celula, Funcao, User
 
 @pytest.fixture
 def client(app_ctx):
